# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: PortalDots API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080/v1
paths:
...
  /contact:
    get:
      operationId: getContactHistory
      summary: Return contact submissions for the current user and circle context
      responses:
        "200":
          description: Contact submission history
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [id, categoryId, categoryName, subject, status, createdAt]
                  properties:
                    id:
                      type: string
                    categoryId:
                      type: string
                    categoryName:
                      type: string
                    subject:
                      type: string
                    status:
                      type: string
                    createdAt:
                      type: string
                      format: date-time
        "401":
          description: Unauthenticated
        "409":
          description: Current circle is not selected
    post:
      operationId: postContact
      summary: Submit a contact message for the current circle context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, subject, body]
              properties:
                categoryId:
                  type: string
                subject:
                  type: string
                  minLength: 1
                body:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Submitted contact
          content:
            application/json:
              schema:
                type: object
                required: [id, categoryId, categoryName, subject, status, createdAt]
                properties:
                  id:
                    type: string
                  categoryId:
                    type: string
                  categoryName:
                    type: string
                  subject:
                    type: string
                  status:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        "401":
          description: Unauthenticated
        "409":
          description: Current circle is not selected
...
  /staff/forms/{formID}/answers/{answerID}/uploads/{questionID}/file:
    get:
      operationId: getStaffFormAnswerUploadFile
      summary: Download an uploaded file for a specific answer question
      parameters:
        - in: path
          name: formID
          required: true
          schema:
            type: string
        - in: path
          name: answerID
          required: true
          schema:
            type: string
        - in: path
          name: questionID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploaded file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /staff/forms/{formID}/uploads/{uploadID}/file:
    get:
      operationId: downloadStaffFormUpload
      summary: Download an uploaded file attached to the current circle answer for a staff form
      parameters:
        - in: path
          name: formID
          required: true
          schema:
            type: string
        - in: path
          name: uploadID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploaded file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification is required
        "404":
          description: Form or upload not found
...
  /forms/{formID}/answers/{answerID}/uploads/{questionID}/file:
    get:
      operationId: downloadFormAnswerUploadByAnswer
      summary: Download a file attached to one answer for the current circle and form
      parameters:
        - in: path
          name: formID
          required: true
          schema:
            type: string
        - in: path
          name: answerID
          required: true
          schema:
            type: string
        - in: path
          name: questionID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploaded file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthenticated
        "404":
          description: Form, answer, or upload not found
        "409":
          description: Current circle is not selected
        "422":
          description: Validation error
...
  /forms/{formID}/answer/uploads/{uploadID}/file:
    get:
      operationId: downloadFormAnswerUpload
      summary: Download a file attached to the current circle answer for a form
      parameters:
        - in: path
          name: formID
          required: true
          schema:
            type: string
        - in: path
          name: uploadID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploaded file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthenticated
        "404":
          description: Form or upload not found
        "409":
          description: Current circle is not selected
        "422":
          description: Validation error
...
  /session/profile:
    put:
      operationId: putSessionProfile
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName]
              properties:
                displayName:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                type: object
                required: [id, displayName]
                properties:
                  id:
                    type: string
                  displayName:
                    type: string
        "401":
          description: Unauthenticated
        "422":
          description: Validation error
...